Node Removed, Database connection added
Removed node.js as it does not seem to be required for what I am currently attempting to do. I have left some code for the connection still in there in case I revisit it however it does not seem to be useable for the browser aspect of the site.

Connection has been added in php along with placeholder data to show that it is being connected to successfully. I have roughed the cards in as well to allow for some sort of structure when I go back through and add interactivity with the cards.