Cards have been added
Roughed in the outline of cards using js models that get passed encoded php data to constructor the models.

General placement has been made for how they are displayed on the page.

Some includes/scripts have been adjusted to be useable in a more generic way. This includes the shared head not including tags to allow for page specific includes/scripts to be added without potential of leaving floating tags.