CSS Changes
Cards now look a bit neater.
They will stack properly and move around when the screen size reduces.
Changed the name for the cards from cardLayout to just card as I am planning on using the class name for other interactive functions soon and the name better represents the display as a whole.
Added a few new entries to the db however it is not reflected inside the backup sql file.