Admin Page added, Changes to Card
There has been a placeholder generated for the admin page and a rough connection to an adminTools.php that will handle form submissions  to the database.

Added the ability for the images to be populated inside the Company page to establish that to work.

Clicking a card now makes the size double, no other changes have been made so far.