Adjustments made for styling
Have adjusted the cards to let them cut off text before it can overflow the card and given it a property to 'open up' to give all the details. This is performed through toggling a hidden class.

Made a few changes within the model as well to allow for the changes to be immediately available and not need to be adjusted during the running of the main script.

The main stylesheet has also had a general max and min size implemented to allow for images to look neater and not break other elements on the page.