Admin Page Updates
Admin page now contains all information on the website and organises it onto tables separated by tabs into their respective areas. When hovered over it will change colour to indicate what entry is going to be selected through a click event.

Cleaned up some bugs that I didn't notice in the data with the information not being collected correctly.

Created an address model and the functions that all the other models have in the php and JS files.

Inline styling added for the images that exist to ensure that no breaking of the table occurs.