Forms started
Added a rough get set up to move from one page to another, will need to refined further in order to get the selected table and pass that information as well otherwise some SQL magic will need to happen to check all tables for the correct name which might not be ideal.

May have to redesign the table slightly to accommodate the changes.