Form additions, slight refactor and comments
Cookies now exist to allow for the information to be passed between pages without requiring POSTs or GETs and to ensure current data is used to populate a form.'

The form is populated but not connected to the backend to allow for updating or creation.

Comments have been added to allow for the removal of aspects that are not used anymore and/or superseded.