Company entries can be updated
Framework for updating of entries has been added.
Updating a company works.
Minor issues seem to be happening from special characters when attempting to print to validate, this does not seem to occur when submission to the database happens.
Cleaning up the switch case statements will need to happen to reduce bloat.